refactor(chat): split Chat component into smaller, focused components
- Extract MessageInput component for input handling
- Extract MessageList component with recipe management
- Create InventoryItemBadge component for reusable item display
- Reduce Chat component from 350+ to ~190 lines
- Improve separation of concerns and maintainability
